requirements.txt
APScheduler==3.3.1
boto3==1.4.4
botocore==1.5.46
docutils==0.13.1
Faker==0.7.11
jmespath==0.9.2
numpy==1.12.1
python-dateutil==2.6.0
s3transfer==0.1.10
tzlocal==1.4